backend/requirements.txt
# Core dependencies
fastapi>=0.115.12
pydantic>=2.0.0
pydantic-settings>=2.2.0
uvicorn>=0.34.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.0
alembic>=1.15.2
python-dotenv>=1.0.0
tenacity>=8.2.0  # For retry logic

# Testing
pytest>=7.4.0
httpx>=0.23.0

# AI dependencies
# Commenting out heavy dependencies for testing purposes
# spacy==3.5.0
# sentence-transformers==2.2.2
# huggingface-hub==0.16.4
openai==1.3.0

# Additional dependencies
typing-extensions==4.13.1

# PDF and DOCX export dependencies
pdfkit==1.0.0
python-docx==0.8.11
jinja2>=3.1.2
email-validator>=2.1.0

# Resume parsing dependencies
pdfplumber==0.7.1
# Removed textract due to dependency conflict with pdfplumber
striprtf==0.0.29
# Alternative for DOC files
docx2txt==0.8

# Payment processing
stripe==5.4.0

# Added from the code block
psycopg2-binary>=2.9.9
pandas==2.2.3
matplotlib==3.10.1
requests==2.32.3
starlette==0.46.1
click==8.1.8

# New dependencies
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
